+++ tests/test_swagger.py
@@ -2756,6 +2756,28 @@ class SwaggerTests(ApiMixin, TestCase):
         self.assertIn('deprecated', post_operation)
         self.assertTrue(post_operation['deprecated'])
 
+    def test_method_restrictions(self):
+        api = self.build_api()
+
+        @api.route('/foo/bar', endpoint='foo')
+        @api.route('/bar', methods=['GET'], endpoint='bar')
+        class TestResource(restplus.Resource):
+            def get(self):
+                pass
+
+            def post(self):
+                pass
+
+        data = self.get_specs()
+
+        path = data['paths']['/foo/bar']
+        self.assertIn('get', path)
+        self.assertIn('post', path)
+
+        path = data['paths']['/bar']
+        self.assertIn('get', path)
+        self.assertNotIn('post', path)
+
 
 class SwaggerDeprecatedTest(ApiMixin, TestCase):
     def test_doc_parser_parameters(self):